# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Streetly API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change the current user's password (also clears the must-change-password flag)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
...
  /agents/apply:
    post:
      operationId: applyAsAgent
      tags: [agents]
      summary: Apply to become a field agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{agentId}/dashboard:
    get:
      operationId: getAgentDashboard
      tags: [agents]
      summary: Get agent dashboard stats
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDashboard"
  /agents/{agentId}/listings:
    get:
      operationId: getAgentListings
      tags: [agents]
      summary: Get businesses submitted by agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Business"
  /agents/{agentId}/withdraw:
    post:
      operationId: requestWithdrawal
      tags: [agents]
      summary: Request earnings withdrawal
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Withdrawal requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [visitor, business_owner, field_agent]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string

    # Contact
...
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
        mustChangePassword:
          type: boolean

    # Stats
...
    BusinessPhoto:
      type: object
      required: [id, url]
      properties:
        id:
          type: integer
        url:
          type: string
        caption:
          type: ["string", "null"]
    Business:
      type: object
      required: [id, name, categoryId, streetId, status, createdAt]
      properties:
        id:
          type: integer
        slug:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        streetId:
          type: integer
        streetName:
          type: ["string", "null"]
        areaName:
          type: ["string", "null"]
        cityName:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        openingHours:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, suspended]
        verified:
          type: boolean
        featured:
          type: boolean
        plan:
          type: string
          enum: [free, premium]
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        photos:
          type: array
          items:
            $ref: "#/components/schemas/BusinessPhoto"
        ownerId:
          type: ["integer", "null"]
        agentId:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    Agent:
      type: object
      required: [id, userId, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userEmail:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, suspended]
        totalListings:
          type: integer
        approvedListings:
          type: integer
        totalEarnings:
          type: number
        availableBalance:
          type: number
        bankName:
          type: ["string", "null"]
        accountNumber:
          type: ["string", "null"]
        accountName:
          type: ["string", "null"]
        createdAt:
          type: string
    AgentApplicationInput:
      type: object
      required: [bankName, accountNumber, accountName]
      properties:
        bankName:
          type: string
        accountNumber:
          type: string
        accountName:
          type: string
        idType:
          type: string
        idNumber:
          type: string
    AgentDashboard:
      type: object
      required: [agent, recentListings]
      properties:
        agent:
          $ref: "#/components/schemas/Agent"
        recentListings:
          type: array
          items:
            $ref: "#/components/schemas/Business"
        recentWithdrawals:
          type: array
          items:
            $ref: "#/components/schemas/Withdrawal"
    Withdrawal:
      type: object
      required: [id, agentId, amount, status, createdAt]
      properties:
        id:
          type: integer
        agentId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending, processing, completed, failed]
        createdAt:
          type: string
    WithdrawalInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
...